# This removes the FetchedValue wrapping by default