--- a/src/main/java/graphql/execution/instrumentation/parameters/InstrumentationFieldCompleteParameters.java
+++ b/src/main/java/graphql/execution/instrumentation/parameters/InstrumentationFieldCompleteParameters.java
@@ -48,7 +48,13 @@ public class InstrumentationFieldCompleteParameters {
         return executionStepInfo.get();
     }
 
-    public Object getFetchedValue() {
+    /**
+     * This returns the object that was fetched, ready to be completed as a value.  This can sometimes be a {@link graphql.execution.FetchedValue} object
+     * but most often it's a simple POJO.
+     *
+     * @return the object was fetched read
+     */
+    public Object getFetchedObject() {
         return fetchedValue;
     }