--- a/src/test/groovy/graphql/execution/BreadthFirstExecutionTestStrategy.java
+++ b/src/test/groovy/graphql/execution/BreadthFirstExecutionTestStrategy.java
@@ -22,11 +22,11 @@ public class BreadthFirstExecutionTestStrategy extends ExecutionStrategy {
     public CompletableFuture<ExecutionResult> execute(ExecutionContext executionContext, ExecutionStrategyParameters parameters) throws NonNullableFieldWasNullException {
         MergedSelectionSet fields = parameters.getFields();
 
-        Map<String, FetchedValue> fetchedValues = new LinkedHashMap<>();
+        Map<String, Object> fetchedValues = new LinkedHashMap<>();
 
         // first fetch every value
         for (String fieldName : fields.keySet()) {
-            FetchedValue fetchedValue = fetchField(executionContext, parameters, fields, fieldName);
+            Object fetchedValue = fetchField(executionContext, parameters, fields, fieldName);
             fetchedValues.put(fieldName, fetchedValue);
         }
 
@@ -34,7 +34,7 @@ public class BreadthFirstExecutionTestStrategy extends ExecutionStrategy {
         Map<String, Object> results = new LinkedHashMap<>();
         for (String fieldName : fetchedValues.keySet()) {
             MergedField currentField = fields.getSubField(fieldName);
-            FetchedValue fetchedValue = fetchedValues.get(fieldName);
+            Object fetchedValue = fetchedValues.get(fieldName);
 
             ResultPath fieldPath = parameters.getPath().segment(fieldName);
             ExecutionStrategyParameters newParameters = parameters
@@ -51,17 +51,17 @@ public class BreadthFirstExecutionTestStrategy extends ExecutionStrategy {
         return CompletableFuture.completedFuture(new ExecutionResultImpl(results, executionContext.getErrors()));
     }
 
-    private FetchedValue fetchField(ExecutionContext executionContext, ExecutionStrategyParameters parameters, MergedSelectionSet fields, String fieldName) {
+    private Object fetchField(ExecutionContext executionContext, ExecutionStrategyParameters parameters, MergedSelectionSet fields, String fieldName) {
         MergedField currentField = fields.getSubField(fieldName);
 
         ResultPath fieldPath = parameters.getPath().segment(fieldName);
         ExecutionStrategyParameters newParameters = parameters
                 .transform(builder -> builder.field(currentField).path(fieldPath));
 
-        return Async.<FetchedValue>toCompletableFuture(fetchField(executionContext, newParameters)).join();
+        return Async.toCompletableFuture(fetchField(executionContext, newParameters)).join();
     }
 
-    private void completeValue(ExecutionContext executionContext, Map<String, Object> results, String fieldName, FetchedValue fetchedValue, ExecutionStrategyParameters newParameters) {
+    private void completeValue(ExecutionContext executionContext, Map<String, Object> results, String fieldName, Object fetchedValue, ExecutionStrategyParameters newParameters) {
         Object resolvedResult = completeField(executionContext, newParameters, fetchedValue).getFieldValueFuture().join();
         results.put(fieldName, resolvedResult);
     }