related to https://github.com/graphql-java/graphql-java/issues/3922

This removes `FetchedValue` wrapping of every value we get back from a DataFetcher

This only wraps it in a FetchedValue if the DF returned the complex `graphql.execution.DataFetcherResult`.

The consuming code now needs to tests if its a `FetchedValue` to get its underlying value or to use it direct