@bbakerman 2025-04-23
Manually run JMH tests

```

This branch

Benchmark                                                                           Mode  Cnt            Score            Error   Units
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesThroughput                     thrpt    4            0.247 ±          0.038   ops/s # better
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesThroughput:gc.alloc.rate       thrpt    4         4076.203 ±        258.101  MB/sec # better
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesThroughput:gc.alloc.rate.norm  thrpt    4  17320751492.000 ± 2385254881.326    B/op # slightly better
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesThroughput:gc.count            thrpt    4           69.000                   counts # same
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesThroughput:gc.time             thrpt    4        11174.000                       ms # better
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesAvgTime                         avgt    4            4.228 ±          1.047    s/op # worse
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesAvgTime:gc.alloc.rate           avgt    4         4019.164 ±        992.876  MB/sec # better
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesAvgTime:gc.alloc.rate.norm      avgt    4  17800984633.333 ±    7488957.877    B/op # better
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesAvgTime:gc.count                avgt    4           80.000                   counts # worse
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesAvgTime:gc.time                 avgt    4        11929.000                       ms # worse


master


Benchmark                                                                           Mode  Cnt            Score            Error   Units
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesThroughput                     thrpt    4            0.260 ±          0.049   ops/s
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesThroughput:gc.alloc.rate       thrpt    4         4382.516 ±        694.460  MB/sec
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesThroughput:gc.alloc.rate.norm  thrpt    4  17680353848.000 ± 1492524470.152    B/op
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesThroughput:gc.count            thrpt    4           72.000                   counts
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesThroughput:gc.time             thrpt    4         8996.000                       ms
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesAvgTime                         avgt    4            3.904 ±          1.378    s/op
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesAvgTime:gc.alloc.rate           avgt    4         4439.397 ±       1687.989  MB/sec
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesAvgTime:gc.alloc.rate.norm      avgt    4  18133700842.667 ± 2305552942.125    B/op
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesAvgTime:gc.count                avgt    4           70.000                   counts
LargeInMemoryQueryPerformance.benchMarkSimpleQueriesAvgTime:gc.time                 avgt    4         9392.000                       ms

```

It definitely helped in using less memory.  Interestingly the gc.time went up - not sure how to read that